+++ R/xl_open.R
@@ -1,23 +1,26 @@
-#' @name xl_open
-#' @title Open a Microsoft Excel file (xls/xlsx) or an openxlsx Workbook
-#' @description This function tries to open a Microsoft Excel
-#' (xls/xlsx) file or an openxlsx Workbook with the proper
-#' application, in a portable manner.
+#' Open a Microsoft Excel file (xls/xlsx) or an openxlsx2 wbWorkbook
 #'
-#' In Windows (c) and Mac (c), it uses system default handlers,
-#' given the file type.
+#' @description This function tries to open a Microsoft Excel (xls/xlsx) file or
+#'   an openxlsx2 wbWorkbook with the proper application, in a portable manner.
 #'
-#' In Linux it searches (via `which`) for available xls/xlsx
-#' reader applications (unless `options('openxlsx.excelApp')`
-#' is set to the app bin path), and if it finds anything, sets
-#' `options('openxlsx.excelApp')` to the program chosen by
-#' the user via a menu (if many are present, otherwise it will
-#' set the only available). Currently searched for apps are
-#' Libreoffice/Openoffice (`soffice` bin), Gnumeric
-#' (`gnumeric`) and Calligra Sheets (`calligrasheets`).
+#'   In Windows it uses `base::shell.exec()` (Windwos only function) to
+#'   determine the appropriate program.
 #'
-#' @param file path to the Excel (xls/xlsx) file or Workbook object.
-#' @usage xl_open(file=NULL)
+#'   In Mac (c) it uses system default handlers, given the file type.
+#'
+#'   In Linux it searches (via `which`) for available xls/xlsx reader
+#'   applications (unless `options('openxlsx2.excelApp')` is set to the app bin
+#'   path), and if it finds anything, sets `options('openxlsx2.excelApp')` to the
+#'   program chosen by the user via a menu (if many are present, otherwise it
+#'   will set the only available). Currently searched for apps are
+#'   Libreoffice/Openoffice (`soffice` bin), Gnumeric (`gnumeric`) and Calligra
+#'   Sheets (`calligrasheets`).
+#'
+#' @param x A path to the Excel (xls/xlsx) file or Workbook object.
+#' @param file Deprecated
+#' @param interactive If `FALSE` will throw a warning and not open the path.
+#'   This can be manually set to `TRUE`, otherwise when `NA` (defualt) uses the
+#'   value returned from [base::interactive()]
 #' @export
 #' @examples
 #' # file example
@@ -30,42 +33,59 @@
 #' wb$add_worksheet("Cars")
 #' wb$add_data("Cars", x, startCol = 2, startRow = 3, rowNames = TRUE)
 #' xl_open(wb)
-xl_open <- function(file = NULL) {
+xl_open <- function(x, file, interactive = NA) {
+  if (!missing(file))  {
+    warning("xl_open(file = .) is deprecated.  Use xl_open(x = .) instead")
+    x <- file
+  }
+  UseMethod("xl_open")
+}
 
-  if (!interactive()) return()
+#' @rdname xl_open
+#' @export
+xl_open.wbWorkbook <- function(x, file, interactive = NA) {
+  stopifnot(R6::is.R6(x))
+  xl_open(x$clone()$save(temp_xlsx())$path, interactive = interactive)
+}
 
-  op <- openxlsx_options()
-  on.exit(options(op), add = TRUE)
+#' @rdname xl_open
+#' @export
+xl_open.default <- function(x, file, interactive = NA) {
+  stopifnot(file.exists(x))
 
-  if (is.null(file)) stop("A file has to be specified.")
+  # nocov start
+  if (is.na(interactive)) {
+    interactive <- interactive()
+  }
+  # nocov end
 
-  ## workbook handling
-  if (inherits(file, "wbWorkbook")) {
-    file <- file$save(path = temp_xlsx())$path
+  if (!isTRUE(interactive)) {
+    warning("will not open file when not interactive")
+    return()
   }
 
-  if (!file.exists(file)) stop("Non existent file or wrong path.")
+  # nocov start
 
   ## execution should be in background in order to not block R
   ## interpreter
-  file <- normalizePath(file)
+  file <- normalizePath(x, mustWork = TRUE)
   userSystem <- Sys.info()["sysname"]
 
-
-  if ("Linux" == userSystem) {
-    if (is.null(app <- unlist(options("openxlsx2.excelApp")))) {
-      app <- chooseExcelApp()
-    }
-    myCommand <- paste(app, file, "&", sep = " ")
-    system(command = myCommand)
-  } else if ("Windows" == userSystem) {
-    shell(shQuote(file, "sh"), wait = FALSE) # nolint
-  } else if ("Darwin" == userSystem) {
-    myCommand <- paste0('open ', shQuote(file))
-    system(command = myCommand)
-  } else {
-    warning("Operating system not handled.")
-  }
+  switch(
+    userSystem,
+    Linux = {
+      app <- getOption("openxlsx2.excelApp", chooseExcelApp())
+      system2(app, c(file, "&"))
+    },
+    Windows = {
+      shell.exec(file) # nolint
+    },
+    Darwin = {
+      system2('open', shQuote(file))
+    },
+    stop("Operating system not handled: ", toString(userSystem))
+  )
+  # nocov end
 }
 
 
@@ -78,19 +98,26 @@ chooseExcelApp <- function() {
 
   prog <- Sys.which(m)
   names(prog) <- names(m)
-  nApps <- length(availProg <- prog["" != prog])
+  availProg <- prog["" != prog]
+  nApps <- length(availProg)
 
   if (0 == nApps) {
     stop(
       "No applications (detected) available.\n",
       "Set options('openxlsx.excelApp'), instead."
     )
-  } else if (1 == nApps) {
-    cat("Only", names(availProg), "found; I'll use it.\n")
+  }
+
+  if (1 == nApps) {
+    message("Only ", names(availProg), " found")
     unnprog <- unname(availProg)
-    options(openxlsx.excelApp = unnprog)
-    invisible(unnprog)
-  } else if (1 < nApps) {
+    message(sprintf("Setting options(openxlsx2.excelApp = '%s')", unnprog))
+    options(openxlsx2.excelApp = unnprog)
+    return(invisible(unnprog))
+  }
+
+  # nocov start
+  if (1 < nApps) {
     if (!interactive()) {
       stop(
         "Cannot choose an Excel file opener non-interactively.\n",
@@ -99,9 +126,14 @@ chooseExcelApp <- function() {
     }
     res <- menu(names(availProg), title = "Excel Apps availables")
     unnprog <- unname(availProg[res])
-    if (res > 0L) options(openxlsx.excelApp = unnprog)
-    invisible(unname(unnprog))
-  } else {
-    stop("Unexpected error.")
+    if (res > 0L) {
+      message(sprintf("Setting options(openxlsx2.excelApp = '%s')", unnprog))
+      options(openxlsx2.excelApp = unnprog)
+    }
+    return(invisible(unname(unnprog)))
   }
+  # nocov end
+
+  stop("Unexpected error in openxlsx2::chooseExcelApp()") # nocov
+
 }